Campaign: Accelerate phases, MIL fix, test infra repair, Residency/Metal wiring
- PhaseGraphBuilder emits AccelerateBlock (ResidualRmsNorm) phases between
  LayerAttention and LayerMlp with primary_lane "accelerate"
- WeightResidencyRunner logs weight staging info
- ResidualRmsNormRunner calls AccelerateLane::rms_norm() + add()
- MIL builder const_f32 auto-fills zero data for empty values with non-empty
  shape (fixes coremlcompiler shape mismatch)
- All 85 pre-existing test compilation errors fixed across 20+ files
- build_rope_tables made pub(crate) for test access
- RealCheckpointReference serde import in test module
- exo.rs borrow fix, executable_bindings.rs buf lifetime fix
- AudioToAudioGenerator Debug derive
- live_execution_matrix.rs comparison operator parentheses

+++ compute-native/compute-core/src/backend/flex_dispatch.rs
@@ -772,7 +772,7 @@ mod tests {
         // resets. We verify the internal state by checking the step counter.
         for i in 0..4 {
             let op = make_matmul_op(i);
-            flex.dispatch(&op, i);
+            flex.dispatch(&op, i as u32);
             assert!(
                 flex.steps_since_sample <= 5,
                 "steps_since_sample should be <= 5 after {i} dispatches"
